Split all the node types into different files
